--- users/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import CustomUser
-
-@admin.register(CustomUser)
-class CustomUserAdmin(admin.ModelAdmin):
-    list_display = ["username", "email", "phone_number", "avatar"]
-# Register your models here.
